moved dataset utils to notebooks
