+++ notebooks/test_utils/utils_mnli.py
@@ -12,7 +12,7 @@
 """
 import os
 import pandas as pd
-from interpret_text.common.dataset.utils_data_shared import maybe_download, extract_zip
+from notebooks.test_utils.utils_data_shared import maybe_download, extract_zip
 
 URL = "http://www.nyu.edu/projects/bowman/multinli/multinli_1.0.zip"
 DATA_FILES = {
@@ -25,7 +25,7 @@ DATA_FILES = {
 def download_file_and_extract(
     local_cache_path: str = ".", file_split: str = "train"
 ) -> None:
-    """Download and extract the dataset files
+    """Download and extract the test_utils files
 
     Args:
         local_cache_path (str [optional]) -- Directory to cache files to.
@@ -39,8 +39,8 @@ def download_file_and_extract(
         extract_zip(os.path.join(local_cache_path, file_name), local_cache_path)
 
 
-def load_pandas_df(local_cache_path=".", file_split="train"):
-    """Loads extracted dataset into pandas
+def load_mnli_pandas_df(local_cache_path=".", file_split="train"):
+    """Loads extracted test_utils into pandas
     Args:
         local_cache_path ([type], optional): [description].
             Defaults to current working directory.
