+++ test/conftest.py
@@ -1,7 +1,9 @@
 import os
+import sys
 import pytest
 from tempfile import TemporaryDirectory
 
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
 def path_notebooks():
     """Returns the path of the notebooks folder"""
